Force garbage collection to prevent memory growth
This is a long running process and the memory started to grow due to unreferenced objects not being removed with the system's garbage collection.  Forcing the GC makes the script solid with no memory growth.  More work could be done to make sure objects are deleted to possibly speed up garbage collection when they go out of scope due to function returns.